requirements.txt
absl-py==0.2.0
analyse==1.0.0
appnope==0.1.0
astor==0.6.2
astroid==1.6.3
aubio==0.4.6
audioread==2.1.5
autopep8==1.3.5
backcall==0.1.0
bleach==1.5.0
brewer2mpl==1.4.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
comet-ml==1.0.8
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
findspark==1.2.0
Flask==1.0.2
Flask-Cors==3.0.4
fonttools==3.25.0
fuzzywuzzy==0.16.0
gast==0.2.0
ggplot==0.11.5
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
kaggle==1.3.1
Keras==2.1.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
librosa==0.6.0
llvmlite==0.22.0
mahotas==1.4.4
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
music21==5.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numba==0.37.0
numpy==1.14.2
opencv-contrib-python==3.4.0.12
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.2.0
patsy==0.5.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
progress==1.3
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pycodestyle==2.4.0
pycparser==2.18
Pygments==2.2.0
pylint==1.8.4
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
resampy==0.2.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sounddevice==0.3.11
SpeechRecognition==3.8.1
statsmodels==0.8.0
tensorboard==1.7.0
tensorflow==1.7.0
tensorflow-gpu==1.1.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tflearn==0.3.2
torch==0.4.0
torchvision==0.2.1
tornado==5.0.2
tqdm==4.23.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.47.0
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
wurlitzer==1.0.1
